requirements.txt
Flask==3.1.0
numpy==2.0.2
opencv-python==4.11.0.86
pillow==11.1.0
scikit-image==0.24.0
scipy==1.13.1
tensorflow==2.19.0
torch==2.6.0
torchvision==0.21.0
transformers==4.49.0
matplotlib >=3.3
